# PERF: inline small functions