@ev-br 2021-03-20
On a longer, run the profile's more sane:

```
$ python -m yep -- ViralSim.py example/example.rt -it 70000000 -pm example/example.pp example/example.mg
Total number of iterations:  62277879
Number of lineages of each hyplotype: 16  16  16  16  
popNum:  4
dim:  2
hapNum:  16
totalRate:  0.0
rn:  0.48484767254134337
susceptible:  731263
lbCounter:  0
Current time:  3.3476571289584665
Tree size:  6155381
Number of sampled elements:  3077691
11.861364841461182
13.560812711715698
_________________________________
PROFILE: interrupts/evictions/bytes = 2572/2082/524384
(mc_lib_2) br@duneyrr:~/sweethome/proj/ViralSim$ google-pprof --text --lines `which python` ViralSim.py.prof |more
Using local file /home/br/virtualenvs/mc_lib_2/bin/python.
Using local file ViralSim.py.prof.
Total: 2572 samples
     125   4.9%   4.9%      136   5.3% PySet_Contains ??:0
     114   4.4%   9.3%      123   4.8% PyLong_FromSsize_t ??:0
     107   4.2%  13.5%      107   4.2% __pyx_f_16BirthDeathCython_15BirthDeathModel_HapPopRate (inline) /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:7673
     106   4.1%  17.6%      106   4.1% f64xsubf128 ??:0
     102   4.0%  21.5%      102   4.0% __Pyx_GetItemInt_Fast [clone .constprop.0] /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:31761
      90   3.5%  25.0%       90   3.5% __Pyx_INC_MEMVIEW (inline) /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:31626
      90   3.5%  28.5%       90   3.5% __pyx_f_5numpy_6random_6_pcg64_pcg64_double _pcg64.c:0
      62   2.4%  30.9%       62   2.4% __pyx_f_16BirthDeathCython_15BirthDeathModel_SampleTime (inline) /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:8521
      54   2.1%  33.0%       54   2.1% __Pyx_GetItemInt_Fast [clone .constprop.0] (inline) /usr/include/python3.8/object.h:459
      51   2.0%  35.0%       51   2.0% __Pyx_XDEC_MEMVIEW (inline) /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:31649
      50   1.9%  37.0%       50   1.9% __pyx_f_16BirthDeathCython_15BirthDeathModel_Birth /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:8718
      49   1.9%  38.9%       49   1.9% __pyx_f_16BirthDeathCython_15BirthDeathModel_GenerateEvent (inline) /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:31649
      46   1.8%  40.7%      254   9.9% _PyObject_LookupSpecial ??:0
      46   1.8%  42.5%       78   3.0% __pyx_f_16BirthDeathCython_6Events_GetEvent /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:4330
      45   1.7%  44.2%       45   1.7% __pyx_f_16BirthDeathCython_15BirthDeathModel_Birth /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:8707
      40   1.6%  45.8%       40   1.6% PyObject_Malloc ??:0
      36   1.4%  47.2%       51   2.0% __pyx_f_16BirthDeathCython_6Events_GetEvent /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:4318
      33   1.3%  48.4%       33   1.3% _Py_Dealloc ??:0
      32   1.2%  49.7%       32   1.2% __pyx_f_16BirthDeathCython_6Events_AddEvent /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:4262
      31   1.2%  50.9%       31   1.2% __pyx_f_16BirthDeathCython_15BirthDeathModel_Death /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:8934
      29   1.1%  52.0%       29   1.1% __pyx_f_16BirthDeathCython_6Events_AddEvent /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:4252
      29   1.1%  53.1%       29   1.1% __sched_yield ??:0
      26   1.0%  54.2%       26   1.0% PyObject_Free ??:0
      26   1.0%  55.2%       26   1.0% __pyx_f_16BirthDeathCython_15BirthDeathModel_Death /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:8966
      26   1.0%  56.2%       26   1.0% __pyx_f_16BirthDeathCython_15BirthDeathModel_GetGenealogy /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:9680
      25   1.0%  57.2%       25   1.0% __pyx_f_16BirthDeathCython_15BirthDeathModel_BirthRate (inline) /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:7715
      25   1.0%  58.1%       25   1.0% __pyx_f_16BirthDeathCython_6Events_AddEvent /home/br/sweethome/proj/ViralSim/BirthDeathCython.cpp:4212
```

ISTM, removing lists of lists from GetGeneology might bring some more 15-20%. Likely not worth it.